add `CompositionDemo` with grid and interactive sections used on /composition and landing pages
Enhance UI and functionality in various components

- DRY pointer event handlers in `Bond.svelte`
- single <main> tag in `+layout.svelte`, remove those littered throughout the app
- update landing `+page.svelte` to include the new `CompositionDemo` and add molecule files to file picker
