+++ src/routes/(demos)/bohr-atoms/bohr-atoms.md
@@ -1,7 +1,5 @@
 # Bohr Atoms
 
-<main>
-
 This solar-system-like visualization of the elements is known as the Bohr model. It was proposed by Niels Bohr in 1913. It should not be viewed as an accurate picture of reality. Quantum mechanics has shown that electrons are really unlocalized wave functions still centered around the nucleus but with much more complicated shapes determined by their quantum numbers $n$, $l$, $m$ and $s$.
 
 - $n$ is the shell number
@@ -10,8 +8,6 @@ This solar-system-like visualization of the elements is known as the Bohr model.
 - $s$ is the spin
 
 In fact this 2d visualization is a simplification even of the incorrect Bohr model in which electrons really orbit in 3d around the nucleus. Yet this animation gives an intuitive understanding of how electrons are placed into shells and how electron energies decrease with increasing shell number. To be precise, the radius of electron orbitals increases with the square of the shell number (shown here as linear due to page width constraints). Meanwhile, the 'kinetic energy' of the electrons decreases linearly with shell number. The orbital period $T$ is proportional $Z^2 / n^3$, where $Z$ is the atomic number. Shown here is $\sqrt(T) / Z$ (the root of the period scaled by atomic number) as else inner-shell electrons of large atoms would be invisibly fast.
-
-</main>
 
 <style>
   p {
