+++ src/routes/+layout.svelte
@@ -81,6 +81,8 @@
 
 <DemoNav />
 
-{@render children?.()}
+<main>
+  {@render children?.()}
+</main>
 
 <Footer />
