+++ src/routes/[slug]/+page.svelte
@@ -102,161 +102,145 @@
   <meta property="og:title" content={head_title} />
 </svelte:head>
 
-<main>
-  <ElementHeading {element} />
+<ElementHeading {element} />
 
-  {#if (element?.discoverer && !element.discoverer.startsWith(`unknown`)) ||
-      element?.year}
-    <p class="discovery">
-      Discovered
-      {#if element?.discoverer && !element.discoverer.startsWith(`unknown`)}
-        by <strong>{element.discoverer}</strong>
-      {/if}
-      {#if typeof element?.year}
-        in <strong>{element.year}</strong>
-      {/if}
-    </p>
-  {/if}
+{#if (element?.discoverer && !element.discoverer.startsWith(`unknown`)) ||
+    element?.year}
+  <p class="discovery">
+    Discovered
+    {#if element?.discoverer && !element.discoverer.startsWith(`unknown`)}
+      by <strong>{element.discoverer}</strong>
+    {/if}
+    {#if typeof element?.year}
+      in <strong>{element.year}</strong>
+    {/if}
+  </p>
+{/if}
 
-  <form>
-    <PropertySelect minSelect={1} />
-    <ColorScaleSelect bind:value={color_scale} selected={[color_scale]} minSelect={1} />
-  </form>
-  <section class="viz">
-    <ElementPhoto {element} missing_msg={window_width < 900 ? `` : `No image for`} />
+<form>
+  <PropertySelect minSelect={1} />
+  <ColorScaleSelect bind:value={color_scale} selected={[color_scale]} minSelect={1} />
+</form>
+<section class="viz">
+  <ElementPhoto {element} />
 
-    <!-- onmouseleave makes ElementScatter always show current element unless user actively hovers another element -->
-    <ElementScatter
-      y={scatter_plot_values}
-      {y_label}
-      {y_unit}
-      {color_scale}
-      y_lim={[0, null]}
-      onmouseleave={() => (selected.element = element)}
-      style="min-height: min(50vmin, 400px)"
-    />
-  </section>
+  <!-- onmouseleave makes ElementScatter always show current element unless user actively hovers another element -->
+  <ElementScatter
+    y={scatter_plot_values}
+    {y_label}
+    {y_unit}
+    {color_scale}
+    y_lim={[0, null]}
+    onmouseleave={() => (selected.element = element)}
+    style="min-height: min(50vmin, 400px)"
+  />
+</section>
 
-  <p class="summary">{@html element.summary}</p>
+<p class="summary">{@html element.summary}</p>
 
-  <section class="flex-wrap">
-    <PeriodicTable
-      tile_props={{ show_name: false, show_number: false }}
-      show_photo={false}
-      disabled
-      style="width: 100%; max-width: 350px"
-      links="name"
-      active_element={element}
-    />
+<section class="flex-wrap">
+  <PeriodicTable
+    tile_props={{ show_name: false, show_number: false }}
+    show_photo={false}
+    disabled
+    style="width: 100%; max-width: 300px"
+    links="name"
+    active_element={element}
+  />
 
-    <table>
-      <thead>
-        <tr>
-          <th><Icon icon="Circle" />&nbsp;Shell</th>
-          <th><Icon icon="Atom" />&nbsp;Electrons</th>
-          <th><Icon icon="Orbit" />&nbsp;Orbitals</th>
+  <table>
+    <thead>
+      <tr>
+        <th><Icon icon="Circle" />&nbsp;Shell</th>
+        <th><Icon icon="Atom" />&nbsp;Electrons</th>
+        <th><Icon icon="Orbit" />&nbsp;Orbitals</th>
+      </tr>
+    </thead>
+
+    <tbody>
+      {#each element.shells as shell_occu, shell_idx (shell_occu + shell_idx)}
+        {@const shell_orbitals = element.electron_configuration
+          .split(` `)
+          .filter((orbital) => orbital.startsWith(`${shell_idx + 1}`))
+          .map((orbital) =>
+            `${orbital.substring(2)} in ${orbital.substring(0, 2)}`
+          )}
+        <tr
+          onmouseenter={() => (active_shell = shell_idx + 1)}
+          onmouseleave={() => (active_shell = null)}
+        >
+          <td>{shell_idx + 1}</td>
+          <td>{shell_occu}</td>
+          <td>{shell_orbitals.join(` + `)}</td>
         </tr>
-      </thead>
+      {/each}
+    </tbody>
+  </table>
 
-      <tbody>
-        {#each element.shells as shell_occu, shell_idx (shell_occu + shell_idx)}
-          {@const shell_orbitals = element.electron_configuration
-            .split(` `)
-            .filter((orbital) => orbital.startsWith(`${shell_idx + 1}`))
-            .map((orbital) =>
-              `${orbital.substring(2)} in ${orbital.substring(0, 2)}`
-            )}
-          <tr
-            onmouseenter={() => (active_shell = shell_idx + 1)}
-            onmouseleave={() => (active_shell = null)}
-          >
-            <td>{shell_idx + 1}</td>
-            <td>{shell_occu}</td>
-            <td>{shell_orbitals.join(` + `)}</td>
-          </tr>
-        {/each}
-      </tbody>
-    </table>
+  <BohrAtom
+    symbol={element.symbol}
+    shells={element.shells}
+    name={element.name}
+    adapt_size
+    orbital_period={orbiting ? 3 : 0}
+    highlight_shell={active_shell}
+    onclick={() => (orbiting = !orbiting)}
+    style="max-width: 300px"
+  />
+</section>
 
-    <BohrAtom
-      symbol={element.symbol}
-      shells={element.shells}
-      name={element.name}
-      adapt_size
-      orbital_period={orbiting ? 3 : 0}
-      highlight_shell={active_shell}
-      onclick={() => (orbiting = !orbiting)}
-      style="max-width: 300px"
-    />
-  </section>
+<section class="properties">
+  {#each key_vals as [label, value, icon], idx ([label, value, icon])}
+    <!-- skip last item if index is uneven to avoid single dangling item on last row -->
+    {#if idx % 2 === 1 || idx < key_vals.length - 1}
+      <div>
+        <strong>
+          <Icon {icon} />
+          {@html value}
+        </strong>
+        <small>{label}</small>
+      </div>
+    {/if}
+  {/each}
+</section>
 
-  <section class="properties">
-    {#each key_vals as [label, value, icon], idx ([label, value, icon])}
-      <!-- skip last item if index is uneven to avoid single dangling item on last row -->
-      {#if idx % 2 === 1 || idx < key_vals.length - 1}
-        <div>
-          <strong>
-            <Icon {icon} />
-            {@html value}
-          </strong>
-          <small>{label}</small>
-        </div>
-      {/if}
-    {/each}
-  </section>
-
-  <PrevNext
-    items={element_data.map((elem) => [elem.name.toLowerCase(), elem])}
-    current={page.url.pathname.slice(1)}
-  >
-    {#snippet children({ item, kind })}
-      {@const element = item[1] as ChemicalElement}
-      <a
-        href={element.name.toLowerCase()}
-        style="display: flex; flex-direction: column; position: relative"
-      >
-        <h3>
-          {@html kind == `next` ? `Next &rarr;` : `&larr; Previous`}
-        </h3>
-        <ElementPhoto {element} style="width: 200px; border-radius: 4pt" />
-        <ElementTile
-          {element}
-          style="width: 70px; position: absolute; bottom: 0"
-          --elem-tile-hover-border="1px solid transparent"
-        />
-      </a>
-    {/snippet}
-  </PrevNext>
-</main>
+<PrevNext
+  items={element_data.map((elem) => [elem.name.toLowerCase(), elem])}
+  current={page.url.pathname.slice(1)}
+>
+  {#snippet children({ item, kind })}
+    {@const element = item[1] as ChemicalElement}
+    <a
+      href={element.name.toLowerCase()}
+      style="display: flex; flex-direction: column; position: relative"
+    >
+      <h3>
+        {@html kind == `next` ? `Next &rarr;` : `&larr; Previous`}
+      </h3>
+      <ElementPhoto {element} style="width: 200px; border-radius: 4pt" />
+      <ElementTile
+        {element}
+        style="width: 70px; position: absolute; bottom: 0"
+        --elem-tile-hover-border="1px solid transparent"
+      />
+    </a>
+  {/snippet}
+</PrevNext>
 
 <style>
-  main {
-    margin: auto;
-    max-width: 75em;
-  }
   section.viz {
     margin: 2em auto;
     display: grid;
-    gap: 2em;
-  }
-  @media (min-width: 700px) {
-    section.viz {
-      grid-template-columns: 1fr 2fr;
-    }
+    gap: 1em;
+    grid-template-columns: 1fr 2fr;
   }
   section.flex-wrap {
     margin: 3em auto;
     display: flex;
     flex-wrap: wrap;
-    gap: 2em;
+    gap: 1em;
     place-items: center;
     place-content: center space-around;
-  }
-  p {
-    text-align: center;
-    max-width: var(--max-text-width);
-    margin: 3em auto;
-    font-weight: 200;
   }
   section.properties {
     display: grid;
