+++ src/routes/contributing/+page.svelte
@@ -2,6 +2,4 @@
   import Contributing from '$root/contributing.md'
 </script>
 
-<main>
-  <Contributing />
-</main>
+<Contributing />
