+++ src/routes/mp-[slug]/MpDetails.svelte
@@ -21,100 +21,98 @@
   let aws_url = $derived(`${mp_bucket}/summary/${mp_id}.json.gz`)
 </script>
 
-<main>
-  <center style="margin: 1em 0">
-    <h1>
-      Materials Explorer - {summary?.formula_pretty}
-      <span style="font-weight: lighter">(<a {href}>{mp_id}</a>)</span>
-    </h1>
+<center style="margin: 1em 0">
+  <h1>
+    Materials Explorer - {summary?.formula_pretty}
+    <span style="font-weight: lighter">(<a {href}>{mp_id}</a>)</span>
+  </h1>
 
-    <input
-      placeholder="Enter MP material ID"
-      bind:value={input_value}
-      onkeydown={async (event) => {
-        if (event.key === `Enter`) {
-          goto(`/${mp_id}`)
-          ;({ summary, similarity, robocrys } = await fetch_material_data(mp_id))
-        }
-      }}
-    />
-    <button
-      onclick={async () => {
+  <input
+    placeholder="Enter MP material ID"
+    bind:value={input_value}
+    onkeydown={async (event) => {
+      if (event.key === `Enter`) {
         goto(`/${mp_id}`)
         ;({ summary, similarity, robocrys } = await fetch_material_data(mp_id))
-      }}
-    >
-      Fetch material
-    </button>
-    <span class="download">
-      <button>Save material summary</button>
-      <div>
-        <button
-          onclick={() => {
-            if (!summary) return alert(`No data to download`)
-            download(
-              JSON.stringify(summary, null, 2),
-              `${mp_id}.json`,
-              `application/json`,
-            )
-          }}
-        >
-          JSON
-        </button>
-        <button
-          onclick={async () => {
-            const blob = await fetch_zipped(aws_url, { unzip: false })
-            if (!blob) return
-            download(blob, `${mp_id}.json.gz`, `application/gzip`)
-          }}
-        >
-          Zipped JSON
-        </button>
-      </div>
-    </span>
-  </center>
-  {#if summary?.structure}
-    <Structure structure={summary.structure} />
-  {/if}
-
-  {#if summary}
-    <MaterialCard material={summary} />
-  {/if}
-
-  {#if robocrys?.description}
-    <h3>
-      Crystal description
-      <small>
-        (generated by <a
-          href="https://github.com/hackingmaterials/robocrystallographer"
-        >Robocrystallographer
-        </a>)
-      </small>
-    </h3>
-    <p>{robocrys.description}</p>
-  {/if}
-
-  {#if similarity}
-    <h2>Similar structures</h2>
-    <ol class="similar-structures">
-      {#each similarity.sim?.slice(0, 6) ?? [] as
-        { task_id, dissimilarity, formula }
-        (task_id + formula + dissimilarity)
       }
-        <li>
-          <strong>
-            <a href="https://materialsproject.org/tasks/{task_id}">{task_id}</a>
-          </strong>
-          <span>{formula}</span>
-          <br />
-          <small>dissimilarity: {format_num(dissimilarity)}</small>
-        </li>
-      {:else}
-        <li>No similar structures found</li>
-      {/each}
-    </ol>
-  {/if}
-</main>
+    }}
+  />
+  <button
+    onclick={async () => {
+      goto(`/${mp_id}`)
+      ;({ summary, similarity, robocrys } = await fetch_material_data(mp_id))
+    }}
+  >
+    Fetch material
+  </button>
+  <span class="download">
+    <button>Save material summary</button>
+    <div>
+      <button
+        onclick={() => {
+          if (!summary) return alert(`No data to download`)
+          download(
+            JSON.stringify(summary, null, 2),
+            `${mp_id}.json`,
+            `application/json`,
+          )
+        }}
+      >
+        JSON
+      </button>
+      <button
+        onclick={async () => {
+          const blob = await fetch_zipped(aws_url, { unzip: false })
+          if (!blob) return
+          download(blob, `${mp_id}.json.gz`, `application/gzip`)
+        }}
+      >
+        Zipped JSON
+      </button>
+    </div>
+  </span>
+</center>
+{#if summary?.structure}
+  <Structure structure={summary.structure} />
+{/if}
+
+{#if summary}
+  <MaterialCard material={summary} />
+{/if}
+
+{#if robocrys?.description}
+  <h3>
+    Crystal description
+    <small>
+      (generated by <a
+        href="https://github.com/hackingmaterials/robocrystallographer"
+      >Robocrystallographer
+      </a>)
+    </small>
+  </h3>
+  <p>{robocrys.description}</p>
+{/if}
+
+{#if similarity}
+  <h2>Similar structures</h2>
+  <ol class="similar-structures">
+    {#each similarity.sim?.slice(0, 6) ?? [] as
+      { task_id, dissimilarity, formula }
+      (task_id + formula + dissimilarity)
+    }
+      <li>
+        <strong>
+          <a href="https://materialsproject.org/tasks/{task_id}">{task_id}</a>
+        </strong>
+        <span>{formula}</span>
+        <br />
+        <small>dissimilarity: {format_num(dissimilarity)}</small>
+      </li>
+    {:else}
+      <li>No similar structures found</li>
+    {/each}
+  </ol>
+{/if}
 
 <style>
   .download {
