+++ src/site/index.ts
@@ -1,3 +1,4 @@
+export { default as CompositionDemo } from './CompositionDemo.svelte'
 export { default as DemoNav } from './DemoNav.svelte'
 export { default as FilePicker } from './FilePicker.svelte'
 export { default as Footer } from './Footer.svelte'
