+++ tests/playwright/structure.test.ts
@@ -1,7 +1,7 @@
 // deno-lint-ignore-file no-await-in-loop
-import { DEFAULTS } from '$lib'
 import { expect, type Page, test } from '@playwright/test'
 import { Buffer } from 'node:buffer'
+import { DEFAULTS } from '../../src/lib/settings'
 import { open_structure_controls_panel } from './helpers'
 
 const default_cam_projection = DEFAULTS.structure.projection
